// src/api/ambulance-wl/apis/DepartmentsApi.ts
export interface GetDepartmentRequest {
    departmentId: string;
}

export interface UpdateDepartmentRequest {
    departmentId: string;
    department: Department;
}

/**
 * DepartmentsApi - interface
 * 
 * @export
 * @interface DepartmentsApiInterface
 */
export interface DepartmentsApiInterface {
    /**
     * Returns one department by id.
     * @summary Get department by id
     * @param {string} departmentId Department identifier
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DepartmentsApiInterface
     */
    getDepartmentRaw(requestParameters: GetDepartmentRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<Department>>;

    /**
     * Returns one department by id.
     * Get department by id
     */
    getDepartment(requestParameters: GetDepartmentRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<Department>;

    /**
     * Returns all departments with room configuration.
     * @summary Get all departments
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DepartmentsApiInterface
     */
    getDepartmentsRaw(initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<Array<Department>>>;

    /**
     * Returns all departments with room configuration.
     * Get all departments
     */
    getDepartments(initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<Array<Department>>;

    /**
     * Updates department and its rooms.
     * @summary Update department
     * @param {string} departmentId Department identifier
     * @param {Department} department Full department object to update
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DepartmentsApiInterface
     */
    updateDepartmentRaw(requestParameters: UpdateDepartmentRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<Department>>;

    /**
     * Updates department and its rooms.
     * Update department
     */
    updateDepartment(requestParameters: UpdateDepartmentRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<Department>;

}

/**
 * 
 */
export class DepartmentsApi extends runtime.BaseAPI implements DepartmentsApiInterface {

    /**
     * Returns one department by id.
     * Get department by id
     */
    async getDepartmentRaw(requestParameters: GetDepartmentRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<Department>> {
        if (requestParameters.departmentId === null || requestParameters.departmentId === undefined) {
            throw new runtime.RequiredError('departmentId','Required parameter requestParameters.departmentId was null or undefined when calling getDepartment.');
        }

        const queryParameters: any = {};

        const headerParameters: runtime.HTTPHeaders = {};

        const response = await this.request({
            path: `/departments/{departmentId}`.replace(`{${"departmentId"}}`, encodeURIComponent(String(requestParameters.departmentId))),
            method: 'GET',
            headers: headerParameters,
            query: queryParameters,
        }, initOverrides);

        return new runtime.JSONApiResponse(response, (jsonValue) => DepartmentFromJSON(jsonValue));
    }

    /**
     * Returns one department by id.
     * Get department by id
     */
    async getDepartment(requestParameters: GetDepartmentRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<Department> {
        const response = await this.getDepartmentRaw(requestParameters, initOverrides);
        return await response.value();
    }

    /**
     * Returns all departments with room configuration.
     * Get all departments
     */
    async getDepartmentsRaw(initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<Array<Department>>> {
        const queryParameters: any = {};

        const headerParameters: runtime.HTTPHeaders = {};

        const response = await this.request({
            path: `/departments`,
            method: 'GET',
            headers: headerParameters,
            query: queryParameters,
        }, initOverrides);

        return new runtime.JSONApiResponse(response, (jsonValue) => jsonValue.map(DepartmentFromJSON));
    }

    /**
     * Returns all departments with room configuration.
     * Get all departments
     */
    async getDepartments(initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<Array<Department>> {
        const response = await this.getDepartmentsRaw(initOverrides);
        return await response.value();
    }

    /**
     * Updates department and its rooms.
     * Update department
     */
    async updateDepartmentRaw(requestParameters: UpdateDepartmentRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<Department>> {
        if (requestParameters.departmentId === null || requestParameters.departmentId === undefined) {
            throw new runtime.RequiredError('departmentId','Required parameter requestParameters.departmentId was null or undefined when calling updateDepartment.');
        }

        if (requestParameters.department === null || requestParameters.department === undefined) {
            throw new runtime.RequiredError('department','Required parameter requestParameters.department was null or undefined when calling updateDepartment.');
        }

        const queryParameters: any = {};

        const headerParameters: runtime.HTTPHeaders = {};

        headerParameters['Content-Type'] = 'application/json';

        const response = await this.request({
            path: `/departments/{departmentId}`.replace(`{${"departmentId"}}`, encodeURIComponent(String(requestParameters.departmentId))),
            method: 'PUT',
            headers: headerParameters,
            query: queryParameters,
            body: DepartmentToJSON(requestParameters.department),
        }, initOverrides);

        return new runtime.JSONApiResponse(response, (jsonValue) => DepartmentFromJSON(jsonValue));
    }

    /**
     * Updates department and its rooms.
     * Update department
     */
    async updateDepartment(requestParameters: UpdateDepartmentRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<Department> {
        const response = await this.updateDepartmentRaw(requestParameters, initOverrides);
        return await response.value();
    }

}