// src/api/ambulance-wl/models/Department.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Davgus Ambulance API
 * Patient, department and placement management for Davgus ambulance UI
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: your_email@stuba.sk
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { Room } from './Room';
import {
    RoomFromJSON,
    RoomFromJSONTyped,
    RoomToJSON,
} from './Room';

/**
 * 
 * @export
 * @interface Department
 */
export interface Department {
    /**
     * 
     * @type {string}
     * @memberof Department
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Department
     */
    name: string;
    /**
     * 
     * @type {Array<Room>}
     * @memberof Department
     */
    rooms: Array<Room>;
}

/**
 * Check if a given object implements the Department interface.
 */
export function instanceOfDepartment(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "id" in value;
    isInstance = isInstance && "name" in value;
    isInstance = isInstance && "rooms" in value;

    return isInstance;
}

export function DepartmentFromJSON(json: any): Department {
    return DepartmentFromJSONTyped(json, false);
}

export function DepartmentFromJSONTyped(json: any, ignoreDiscriminator: boolean): Department {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'rooms': ((json['rooms'] as Array<any>).map(RoomFromJSON)),
    };
}

export function DepartmentToJSON(value?: Department | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'name': value.name,
        'rooms': ((value.rooms as Array<any>).map(RoomToJSON)),
    };
}