// src/api/ambulance-wl/models/Patient.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Davgus Ambulance API
 * Patient, department and placement management for Davgus ambulance UI
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: your_email@stuba.sk
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Patient
 */
export interface Patient {
    /**
     * 
     * @type {string}
     * @memberof Patient
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Patient
     */
    name: string;
    /**
     * 
     * @type {Date}
     * @memberof Patient
     */
    birthDate: Date;
    /**
     * 
     * @type {string}
     * @memberof Patient
     */
    insuranceId: string;
    /**
     * 
     * @type {string}
     * @memberof Patient
     */
    insuranceCompany: string;
    /**
     * 
     * @type {string}
     * @memberof Patient
     */
    phone: string;
    /**
     * 
     * @type {string}
     * @memberof Patient
     */
    email: string;
    /**
     * 
     * @type {boolean}
     * @memberof Patient
     */
    archived: boolean;
}

/**
 * Check if a given object implements the Patient interface.
 */
export function instanceOfPatient(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "id" in value;
    isInstance = isInstance && "name" in value;
    isInstance = isInstance && "birthDate" in value;
    isInstance = isInstance && "insuranceId" in value;
    isInstance = isInstance && "insuranceCompany" in value;
    isInstance = isInstance && "phone" in value;
    isInstance = isInstance && "email" in value;
    isInstance = isInstance && "archived" in value;

    return isInstance;
}

export function PatientFromJSON(json: any): Patient {
    return PatientFromJSONTyped(json, false);
}

export function PatientFromJSONTyped(json: any, ignoreDiscriminator: boolean): Patient {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'birthDate': (new Date(json['birthDate'])),
        'insuranceId': json['insuranceId'],
        'insuranceCompany': json['insuranceCompany'],
        'phone': json['phone'],
        'email': json['email'],
        'archived': json['archived'],
    };
}

export function PatientToJSON(value?: Patient | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'name': value.name,
        'birthDate': (value.birthDate.toISOString().substr(0,10)),
        'insuranceId': value.insuranceId,
        'insuranceCompany': value.insuranceCompany,
        'phone': value.phone,
        'email': value.email,
        'archived': value.archived,
    };
}