// src/api/ambulance-wl/models/Room.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Davgus Ambulance API
 * Patient, department and placement management for Davgus ambulance UI
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: your_email@stuba.sk
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Room
 */
export interface Room {
    /**
     * 
     * @type {string}
     * @memberof Room
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Room
     */
    number: string;
    /**
     * 
     * @type {number}
     * @memberof Room
     */
    capacity: number;
    /**
     * 
     * @type {string}
     * @memberof Room
     */
    status: RoomStatusEnum;
}


/**
 * @export
 */
export const RoomStatusEnum = {
    Active: 'active',
    Maintenance: 'maintenance',
    Closed: 'closed'
} as const;
export type RoomStatusEnum = typeof RoomStatusEnum[keyof typeof RoomStatusEnum];


/**
 * Check if a given object implements the Room interface.
 */
export function instanceOfRoom(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "id" in value;
    isInstance = isInstance && "number" in value;
    isInstance = isInstance && "capacity" in value;
    isInstance = isInstance && "status" in value;

    return isInstance;
}

export function RoomFromJSON(json: any): Room {
    return RoomFromJSONTyped(json, false);
}

export function RoomFromJSONTyped(json: any, ignoreDiscriminator: boolean): Room {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'number': json['number'],
        'capacity': json['capacity'],
        'status': json['status'],
    };
}

export function RoomToJSON(value?: Room | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'number': value.number,
        'capacity': value.capacity,
        'status': value.status,
    };
}